requirements.txt
llama-index
llama-index-llms-huggingface
llama-index-embeddings-huggingface
transformers
torch
sentence-transformers
python-dotenv